event page completed

+++ web/src/pages/event_page/comment_event.js
@@ -62,6 +62,7 @@ export default function CommentEvent() {
               const {result, data} = await API.postEventComment({memNo, eventIdx, depth, content})
               if (result === 'success') {
                 fetchCommentData()
+                window.scrollTo(0, document.body.scrollHeight)
               }
             }
             if (token.isLogin) {
@@ -111,8 +112,12 @@ export default function CommentEvent() {
                           fetchCommentData()
                         }
                       }
-
-                      DeleteComment(replyIdx, eventIndex)
+                      globalCtx.action.confirm({
+                        msg: '댓글을 삭제하시겠습니까?',
+                        callback: () => {
+                          DeleteComment(replyIdx, eventIndex)
+                        }
+                      })
                     }}>
                     <img src={deleteIcon}></img>
                   </button>
